Add files via upload
In this folder, we take a look at the insertion sort algorithm. However, there's a twist! I've added a class, friend function, dynamic memory allocation, and so fourth. As a refresher, I'll define some of these te1.  Class: A class is a object that has properties. The properties in this case are it's members. The members are attributes and methods. Attributes are variables whereas methods are functions. The attributes are typically under a private attribute and functions are under a public attribute. We can use setters and getters as a means for the user to access data under the private attribute, but at the risk of more syntax.

2. Friend function: A friend function is a nonmember function of the class but must be declared within the class. It's main perk is that it allows us to access members under any type of attribute. 

3. Dynamic memory allocation: This means that we changing memory is quite flexible, especially during the run time. We use the operators new to execute and delete to remove avoid memory leaks.